Reorganization of folder structre for front-end
