+++ frontend/package.json
@@ -14,13 +14,11 @@
     "@types/react-router": "^5.1.11",
     "@types/react-router-dom": "^5.1.7",
     "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "node-sass": "^4.14.1",
     "react": "^17.0.1",
-    "react-bootstrap": "^1.4.3",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
@@ -54,7 +52,7 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.11.9",
-    "@types/mocha": "^8.2.0",
-    "@types/cors": "^2.8.9"
+    "@types/cors": "^2.8.9",
+    "@types/mocha": "^8.2.0"
   }
 }
