+++ frontend/src/App.module.css
@@ -1,0 +1,6 @@
+html {
+    width: 100%;
+    height: 100%;
+    background: url('./assets/background/geometric.jpg') repeat center;
+    background-size: contain;
+}
