--- frontend/src/css/App.module.css
@@ -1,6 +1,0 @@
-html {
-    width: 100%;
-    height: 100%;
-    background: url('../assets/background/geometric.jpg') repeat center;
-    background-size:contain;
-   }
