+++ frontend/src/pages/index.tsx
@@ -1,0 +1,9 @@
+import Home from './home/Home';
+import Login from './login/Login';
+import Welcome from './welcome/Welcome';
+
+export {
+  Home,
+  Login,
+  Welcome
+}
