+++ frontend/src/pages/welcome/WelcomeButtons.tsx
@@ -1,7 +1,6 @@
-import { render } from '@testing-library/react';
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import styles from '../css/WelcomeButtons.module.css'
+import styles from './WelcomeButtons.module.css'
 
 type WelcomeButtonsPage = {
 
@@ -22,7 +21,7 @@ export class WelcomeButtons extends Component<{}, WelcomeButtonsPage>{
                         <p></p>
                     </td>
                     <td>
-                        
+
                     </td>
                 </tr>
             </table>
